Qt: move code from emu_settings to config_adapter
