+++ rpcs3/rpcs3qt/CMakeLists.txt
@@ -6,6 +6,7 @@
 	call_stack_list.cpp
 	cg_disasm_window.cpp
 	cheat_manager.cpp
+	config_adapter.cpp
 	curl_handle.cpp
 	custom_dialog.cpp
 	debugger_frame.cpp
